webapp/requirements.txt
Django==2.0.4
gunicorn==19.9.0
djangorestframework==3.8.2
PyJWT==1.6.4
psycopg2==2.7.5
